+++ include/flatcc/flatcc_json_parser.h
@@ -64,8 +64,8 @@ enum flatcc_json_parser_flags {
     XX(union_vector_length,     "union vector length mismatch")             \
     XX(base64,                  "invalid base64 content")                   \
     XX(base64url,               "invalid base64url content")                \
-    XX(array_underflow,         "fixed size array underflow")               \
-    XX(array_overflow,          "fixed size array overflow")                \
+    XX(array_underflow,         "fixed length array underflow")               \
+    XX(array_overflow,          "fixed length array overflow")                \
     XX(runtime,                 "runtime error")                            \
     XX(not_supported,           "not supported")
 
@@ -164,7 +164,7 @@ static inline const char *flatcc_json_parser_string_end(flatcc_json_parser_t *ct
 }
 
 /*
- * Parse a string as a fixed size char array as `s` with length `n`.
+ * Parse a string as a fixed length char array as `s` with length `n`.
  * and raise errors according to overflow/underflow runtime flags. Zero
  * and truncate as needed. A trailing zero is not inserted if the input
  * is at least the same length as the char array.